% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data_cocoso.R
\name{check_input_data_cocoso}
\alias{check_input_data_cocoso}
\title{Verify if the dataset to be submitted to cocoso function is correctly formatted.}
\usage{
check_input_data_cocoso(md_data)
}
\arguments{
\item{md_data}{Must be a data frame with information about Criteria, Alternatives and Optimal Value. See example_cocoso.}
}
\value{
True if the dataset is correctly formatted and False otherwise.
}
\description{
Verify if the dataset to be submitted to cocoso function is correctly formatted.
}
\examples{
data(example_cocoso)
check_input_data_cocoso(example_cocoso)
}
