% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoso.R
\name{cocoso}
\alias{cocoso}
\title{Implementation of CoCoSo Method}
\usage{
cocoso(md_data, parameter_alpha = 0.5)
}
\arguments{
\item{md_data}{Must be a data frame with information about Criteria, Alternatives and Optimal Value. See example_cocoso.}

\item{parameter_alpha}{Parameter for third aggregation strategies.}
}
\value{
Final ranking of all alternatives from CoCoSo method. In addition, we compute and present all three aggregation strategies.
}
\description{
Implementation of CoCoSo Method
}
\examples{
data(example_cocoso)
cocoso(example_cocoso)

}
