% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cocoso.R
\docType{data}
\name{example_cocoso}
\alias{example_cocoso}
\title{Data example of logistic provider selection problem.}
\format{
A data frame with 10 rows and 6 columns.
}
\source{
{Extract in "A combined compromise solution (CoCoSo) method for multi-criteria decision-making problems", Yazdani, M., Zarate, P., Kazimieras Zavadskas, E., & Turskis, Z. (2019).}
}
\usage{
example_cocoso
}
\description{
Information of decision-making matrix from a logistic provider selection problem including information about criteria, alternatives and weights.
}
\examples{
example_cocoso


}
\keyword{datasets}
