% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_cocoso.R
\name{standard_cocoso}
\alias{standard_cocoso}
\title{Padronization of decision-making matriz by sum.}
\usage{
standard_cocoso(md_data)
}
\arguments{
\item{md_data}{Decision-making matrix.}
}
\value{
A Decision-making matrix normalize.
}
\description{
Padronization of decision-making matriz by sum.
}
\examples{
standard_cocoso(mtcars)
}
